Moved types to new module
(cherry picked from commit 89ee6bc370715b38a0e9152ab4ee646ec901fbf0)
