+++ components/foundation-types/src/main/java/dk.cloudcreate.essentials.components.foundation.types/SubscriberId.java
