+++ components/foundation-types/src/main/java/dk.cloudcreate.essentials.components.foundation.types/TenantId.java
