+++ components/foundation-types/src/main/java/dk/cloudcreate/essentials/components/eventsourced/eventstore/postgresql/types/GlobalEventOrder.java
@@ -16,18 +16,19 @@
 
 package dk.cloudcreate.essentials.components.eventsourced.eventstore.postgresql.types;
 
-import dk.cloudcreate.essentials.components.eventsourced.eventstore.postgresql.EventStore;
-import dk.cloudcreate.essentials.components.eventsourced.eventstore.postgresql.eventstream.*;
-import dk.cloudcreate.essentials.types.*;
+import dk.cloudcreate.essentials.components.eventsourced.eventstore.postgresql.eventstream.AggregateType;
+import dk.cloudcreate.essentials.types.LongRange;
+import dk.cloudcreate.essentials.types.LongType;
 
 import java.util.List;
-import java.util.stream.*;
+import java.util.stream.Collectors;
+import java.util.stream.LongStream;
 
 import static dk.cloudcreate.essentials.shared.FailFast.requireNonNull;
 
 /**
- * The Global Order is a sequential ever-growing number, which that tracks the order in which events have been stored in the {@link EventStore} table
- * across all {@link AggregateEventStream}'s with the same {@link AggregateType}.<br>
+ * The Global Order is a sequential ever-growing number, which that tracks the order in which events have been stored in the EventStore table
+ * across all AggregateEventStreams with the same {@link AggregateType}.<br>
  * The first global-event-order has value 1, since this is the initial value for a Postgresql BIGINT IDENTITY column.
  */
 public class GlobalEventOrder extends LongType<GlobalEventOrder> {
